Moved packages of utils.compress to utils.codec
