+++ core/src/main/java/hivemall/utils/codec/Base91.java
@@ -45,7 +45,7 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package hivemall.utils.compress;
+package hivemall.utils.codec;
 
 import hivemall.utils.io.FastByteArrayOutputStream;
