+++ core/src/main/java/hivemall/utils/codec/CompressionCodec.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.utils.compress;
+package hivemall.utils.codec;
 
 import java.io.Closeable;
 import java.io.IOException;
