+++ core/src/main/java/hivemall/utils/codec/DeflateCodec.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.utils.compress;
+package hivemall.utils.codec;
 
 import static java.util.zip.Deflater.DEFAULT_COMPRESSION;
