+++ core/src/test/java/hivemall/utils/codec/DeflateCodecTest.java
@@ -16,8 +16,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package hivemall.utils.compress;
+package hivemall.utils.codec;
 
+import hivemall.utils.codec.DeflateCodec;
 import hivemall.utils.io.IOUtils;
 
 import java.io.BufferedInputStream;
